Improvements.

+++ include/TriangleManipulator/ShapeManipulator.hpp
@@ -73,54 +73,11 @@ namespace ShapeManipulator {
     inline void handle_intersections(std::vector<PointLocation::Line>& lines, std::function<void(const PointLocation::Line&, const PointLocation::Line&)> callback) {
         std::multiset<Point_Marker> points = std::multiset<Point_Marker>();
         const PointLocation::Line* lines_ptr = lines.data();
-        for (int i = 0, size = lines.size(); i < size; i++) {
+        for (size_t i = 0, size = lines.size(); i < size; i++) {
             const PointLocation::Line& line = *lines_ptr++;
             points.emplace(line.first, i, BEGIN);
             points.emplace(line.second, i, END);
         }
-        // std::set<std::pair<short, short>> beginnings = std::set<std::pair<short, short>>();
-        // for (const Point_Marker& first : points) {
-        //     short line = first.line;
-        //     const std::set<std::pair<short, short>>::iterator& index = beginnings.find({ lines[line].y1, line });
-        //     std::set<std::pair<short, short>>::iterator mutable_index = index;
-        //     if (index != beginnings.end()) {
-        //         if (index != beginnings.begin()) {
-        //             short first = (*(--mutable_index)).second;
-        //             short second = (*(++++mutable_index)).second;
-        //             if (mutable_index != beginnings.end()) {
-        //                 if (intersects(lines[first], lines[second])) {
-        //                     callback(lines[first], lines[second]);
-        //                     // std::cout << "INTERSECTION CAUGHT A: " << first.second << "," << second.second << std::endl;
-        //                 }
-        //             }
-        //         }
-        //         beginnings.erase(index);
-        //     } else {
-        //         std::set<std::pair<short, short>>::iterator index = beginnings.insert({ first.point.y, line }).first;
-        //         // std::cout << "INSERTING: " << line << std::endl;
-        //         // std::cout << "STACK: " << std::endl;
-        //         // for (const std::pair<short, short>& other_line : beginnings) {
-        //         //     std::cout << other_line.second << "," << other_line.first << std::endl;
-        //         // }
-        //         if (index != beginnings.begin()) {
-        //             const std::pair<short, short>& checking = *(--index);
-        //             if (intersects(lines[line], lines[checking.second])) {
-        //                 callback(lines[line], lines[checking.second]);
-        //                 // std::cout << "INTERSECTION CAUGHT B: " << line << "," << checking.second << std::endl;
-        //             }
-        //             index++;
-        //         }
-        //         if (++index != beginnings.end()) {
-        //             std::pair<short, short> checking = *(index);
-        //             if (intersects(lines[line], lines[checking.second])) {
-        //                 callback(lines[line], lines[checking.second]);
-        //                 // std::cout << "INTERSECTION CAUGHT C: " << line << "," << checking.second << std::endl;
-        //             }
-        //         }
-        //     }
-        // }
-
-        // Ironically, this short little piece of code is somehow faster than the commented stuff above...
         std::set<short> beginnings = std::set<short>();
         for (const Point_Marker& first : points) {
             if (first.marker == BEGIN) {
